requirements.txt
absl-py==1.0.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudpickle==2.1.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
fonttools==4.33.3
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
iopath==0.1.9
jsonpickle==2.2.0
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
mock==4.0.3
network==0.1
networks==0.3.7
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.5
Pillow==9.1.1
portalocker==2.4.0
protobuf==4.21.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
yacs==0.1.8
zipp==3.8.0
